requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
gunicorn==21.2.0
